% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recheck.R
\name{recheck}
\alias{recheck}
\title{Reverse Dependency Check}
\usage{
recheck(sourcepkg, which = "strong", preinstall_dependencies = TRUE)
}
\arguments{
\item{sourcepkg}{path or URL to a source package tarball}

\item{which}{passed to \code{tools::package_dependencies}; set to "most" to
also check reverse suggests.}

\item{preinstall_dependencies}{start by installing dependencies for all
packages to be checked.}
}
\description{
Run a reverse dependency check similar to CRAN.
}
