% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{install_recheck_deps}
\alias{install_recheck_deps}
\title{Wrappers to run in CI}
\usage{
install_recheck_deps(path = ".", which = "strong")
}
\arguments{
\item{path}{path to source package}

\item{which}{either be strong or most}
}
\description{
To call in CI
}
